flask-app/requirements.txt
Flask==0.12
flask-restful==0.3.5
mysql-connector-python==8.0.5
flask_cors==3.0.3
